requirements.txt
Flask==1.1.2
Flask-Login==0.5.0
Jinja2==2.11.2
PyGithub==1.54.1
requests==2.25.1
urllib3==1.26.2